+++ server.go
@@ -182,6 +182,6 @@ Loop:
 }
 
 // SetReadLimit sets the maximum size in bytes for a message read from the peer.
-func (s *Server) SetReadLimit(limit int) {
+func (s *Server) SetReadLimit(limit int64) {
 	maxMessageSize = limit
 }
